Manny's Cortex UI work (enemy vitality HUD, stance diamond, inventory flow,
prompt icons, boss theme) over the perf branch. Conflicts: the two UI files
take the branch, the renderer keeps alternate-frame selection reuse, and
the project's UI nodes take the branch.

+++ engine/crates/psx-engine/src/scene.rs
@@ -346,7 +346,7 @@ pub trait Scene {
     /// wrapper consumes the mask once, after simulation, so held input and
     /// render retries cannot replay a sound.
     #[inline]
-    fn take_gameplay_sfx_events(&mut self) -> u8 {
+    fn take_gameplay_sfx_events(&mut self) -> u16 {
         0
     }
